Autenticando as rotas com JWT

+++ server/src/routes/memories.ts
@@ -3,8 +3,17 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request)=>{
+    await request.jwtVerify()
+  })
+  app.get('/memories', async (request) => {
+   
+    const { user } = request
+    
     const memories = await prisma.memory.findMany({
+      where:{
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -19,7 +28,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request,reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -29,6 +38,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+    if(!memory.isPublic && memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
     return memory
   })
   app.post('/memories', async (request) => {
@@ -43,12 +55,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'd00a75f5-99ad-4f55-97a8-8547ebce863b',
+        userId: request.user.sub ,
       },
     })
     return memory
   })
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request,reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -59,6 +71,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     })
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+    if(memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
     const updatedMemory = await prisma.memory.update({
       where: {
         id,
@@ -72,12 +92,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return updatedMemory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request,reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
-    const memory = await prisma.memory.delete({
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+    if(memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+     await prisma.memory.delete({
       where: {
         id,
       },
